requirements.txt.txt

discord.py==2.6.4
requests
